requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2020.12.5
Django==3.2.5
future==0.18.2
gunicorn==20.1.0
line-bot-sdk==1.19.0
mccabe==0.6.1
sqlparse==0.4.1
pyimgur
matplotlib
opencv-python-headless
pymysql
azure==4.0.0
azure-ai-textanalytics
azure-core
beautifulsoup4==4.10.0
jieba==0.42.1
ehownet==3.0.1